M3-186 Refactor : transactional 처리를 위한 클래스 분리
